+++ Composite/FruitInf.java
@@ -3,5 +3,5 @@ package Composite;
 public interface FruitInf {
     public String getName();
 
-    public int getUnitPrice();
+    public double getUnitPrice();
 }
